% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EyeMetGen_Functions.R
\name{balancing}
\alias{balancing}
\title{balancing}
\usage{
balancing(train_data, test_data, kappa = 1, dupSZ = 20, sd = 2023)
}
\arguments{
\item{train_data}{is the train set}

\item{test_data}{is the test set}

\item{kappa}{is the k parameter of smote function}

\item{dupSZ}{is the dup_size parameter of smote function}
}
\value{
list
}
\description{
it makes balancing for a target
}
