% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EyeMetGen_Functions.R
\name{changenameBestAUC}
\alias{changenameBestAUC}
\title{changenameBestAUC}
\usage{
changenameBestAUC(bestAUC.target)
}
\arguments{
\item{bestAUC.target}{is the name of the model to change}
}
\value{
string
}
\description{
get the best model for a target froam a list of models
}
