% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EyeMetGen_Functions.R
\name{Create_bacteria}
\alias{Create_bacteria}
\title{Create_bacteria}
\usage{
Create_bacteria(path)
}
\arguments{
\item{path}{is the path of the bacteria dataset}
}
\value{
dataframe
}
\description{
Create bacteria dataframe
}
