% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EyeMetGen_Functions.R
\name{Create_sizeTab}
\alias{Create_sizeTab}
\title{Create_sizeTab}
\usage{
Create_sizeTab(split_size_list, trg)
}
\arguments{
\item{split_size_list}{is the list of the size (unbalanced/balanced)}

\item{trg}{is the target}
}
\value{
list
}
\description{
create overview of the size table for a specific target
}
