% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EyeMetGen_Functions.R
\name{Create_targets}
\alias{Create_targets}
\title{Create_targets}
\usage{
Create_targets(path)
}
\arguments{
\item{path}{is the path of the targets dataset}
}
\value{
dataframe
}
\description{
Create  targets dataframe
}
