% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EyeMetGen_Functions.R
\name{doplots}
\alias{doplots}
\title{doplots}
\usage{
doplots(rocobj, calls)
}
\arguments{
\item{rocobj}{is a roc object}

\item{calls}{is a list of plot}
}
\description{
draw ROC plot with Confidence Interval
}
