% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EyeMetGen_Functions.R
\name{draw_confusion_matrix}
\alias{draw_confusion_matrix}
\title{draw_confusion_matrix}
\usage{
draw_confusion_matrix(cm, trg)
}
\arguments{
\item{cm}{is a confusionMatrix object}

\item{trg}{is the target}
}
\description{
draw confusion matrix plot
}
