% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EyeMetGen_Functions.R
\name{draw_confusion_matrix_old}
\alias{draw_confusion_matrix_old}
\title{draw_confusion_matrix_old}
\usage{
draw_confusion_matrix_old(cm, pt = NULL)
}
\arguments{
\item{cm}{is a confusionMatrix object}

\item{filenm}{is the name of file}
}
\description{
draw confusion matrix plot
}
