% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EyeMetGen_Functions_alternative.R
\name{extractGlmnetInfo}
\alias{extractGlmnetInfo}
\title{extractGlmnetInfo}
\usage{
extractGlmnetInfo(object)
}
\arguments{
\item{object}{is a List or Vector}
}
\value{
integer
}
\description{
extract Glmnet Info
}
