% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Feature_Selection_Functions.R
\name{FS_Recursive_Elimination}
\alias{FS_Recursive_Elimination}
\title{FS_Recursive_Elimination}
\usage{
FS_Recursive_Elimination(train_data, test_data, target_col)
}
\arguments{
\item{train_data}{is train set}

\item{test_data}{is test set}

\item{target_col}{is target (i.e., outcome)}
}
\value{
list
}
\description{
Feature Selection - Recursive Feature Elimination
}
