% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Feature_Selection_Functions.R
\name{FS_Wilcox}
\alias{FS_Wilcox}
\title{FS_Wilcox}
\usage{
FS_Wilcox(tar.bact.target, trg)
}
\arguments{
\item{tar.bact.target}{is the full dataset (target+bacteria)}

\item{trg}{is the outcome}
}
\value{
vector
}
\description{
wilcox test
}
