% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EyeMetGen_Functions.R
\name{get_bestModel}
\alias{get_bestModel}
\title{get_bestModel}
\usage{
get_bestModel(TargetmodelsAUCToCompare, TargetmodelsToCompare, trg)
}
\arguments{
\item{trg}{is the target}

\item{Listmodels}{is the list of models for the target}

\item{bestModelAUC}{is the list AUC for all the models}
}
\value{
train object
}
\description{
get the best model for a target froam a list of models
}
