% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Feature_Selection_Functions.R
\name{Get_correlated}
\alias{Get_correlated}
\title{Get_correlated}
\usage{
Get_correlated(dataset)
}
\arguments{
\item{dataset}{is the dataset of the attributes with respect to a target}
}
\value{
dataset with the highly Correlated features
}
\description{
remove high correlated attributes (i.e., features).
Many methods perform better if highly correlated attributes are removed
}
