% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EyeMetGen_Functions.R
\name{get_prediction}
\alias{get_prediction}
\title{get_prediction}
\usage{
get_prediction(model_fit, test_data)
}
\arguments{
\item{model_fit}{is the model to test}

\item{test_data}{is the test dataset}
}
\value{
dataframe
}
\description{
get Performance and statistics
}
