% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EyeMetGen_Functions.R
\name{get_split_size_tabs}
\alias{get_split_size_tabs}
\title{get_split_size_tabs}
\usage{
get_split_size_tabs(train_data, test_data, trg)
}
\arguments{
\item{train_data}{is the train data dataframe}

\item{test_data}{are the test data dataframe}

\item{trg}{is a single target}
}
\description{
get the split size list
}
