% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EyeMetGen_Functions.R
\name{get_split_size_tabs_All}
\alias{get_split_size_tabs_All}
\title{get_split_size_tabs_All}
\usage{
get_split_size_tabs_All(
  pathToLoad = "~/MyDir/EyeMetGenomics/data/Split_RC/RC_dataset.rda",
  targets
)
}
\arguments{
\item{pathToLoad}{path}
}
\description{
get the split size list of all targets
}
