% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EyeMetGen_Functions.R
\name{get_target_df}
\alias{get_target_df}
\title{get_target_df}
\usage{
get_target_df(df_tar.bact, trg)
}
\arguments{
\item{trg}{is the targets}
}
\value{
dataframe
}
\description{
get a dataset subset including only a single target
}
