% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EyeMetGen_Functions.R
\name{Get_taxonomy}
\alias{Get_taxonomy}
\title{Get_taxonomy}
\usage{
Get_taxonomy(path)
}
\arguments{
\item{path}{is the path of the bacteria csv file}
}
\value{
dataframe
}
\description{
Create bacteria dataframe
}
