% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildModels_Functions.R
\name{glmnet_model}
\alias{glmnet_model}
\title{glmnet_model}
\usage{
glmnet_model(train_data, ctrl, sd = 2023)
}
\arguments{
\item{train_data}{is train set}

\item{ctrl}{is la list (i.e. returned by TrainControl function)}

\item{sd}{is the seed}
}
\value{
train object
}
\description{
Create  targets dataframe
}
