% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EyeMetGen_Functions.R
\name{merge_preprocessing}
\alias{merge_preprocessing}
\title{merge_preprocessing}
\usage{
merge_preprocessing(targets, bacteria)
}
\arguments{
\item{targets}{is the targets dataset}

\item{bacteria}{is the bacteria dataset}
}
\value{
dataframe
}
\description{
Merge bacteria and targets dataframes
}
