% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EyeMetGen_Functions.R,
%   R/EyeMetGen_Functions_alternative.R
\name{my_finalize_func}
\alias{my_finalize_func}
\title{my_finalize_func}
\usage{
my_finalize_func(
  result_tuning,
  my_recipe,
  my_model,
  q,
  data_in_scope_test,
  rec,
  trg
)

my_finalize_func(
  result_tuning,
  my_recipe,
  my_model,
  q,
  data_in_scope_test,
  rec,
  trg
)
}
\arguments{
\item{result_tuning}{is the tune_grid function output}

\item{my_recipe}{is the preprocessing function output}

\item{my_model}{is the finalize_model function output}
}
\value{
list

dataframe
}
\description{
Function to finalize the recipe and the model and return the AUC value and the ROC curve of the tuned model.

Function to finalize the recip and the model and returne the AUC value and the ROC curve of the tuned model.
}
