% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EyeMetGen_Functions.R,
%   R/EyeMetGen_Functions_alternative.R
\name{nonzeroCoef}
\alias{nonzeroCoef}
\title{nonzeroCoef}
\usage{
nonzeroCoef(beta, bystep = FALSE)

nonzeroCoef(beta, bystep = FALSE)

nonzeroCoef(beta, bystep = FALSE)
}
\arguments{
\item{beta}{is matrix of coefficients}

\item{bystep}{FALSE means which variables were ever nonzero/bystep = TRUE means which variables are nonzero for each step}

\item{model_fit}{is a glmnet model (ridge,lasso,elastic model)}
}
\value{
A numeric vector with the Extracted Non-Zero Model Coefficients

integer

integer
}
\description{
save roc curve

get the non zero coefficients

get the non zero coefficients
}
