% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EyeMetGen_Functions.R
\name{plot_bestModelConfusionMatrix}
\alias{plot_bestModelConfusionMatrix}
\title{plot_bestModelConfusionMatrix}
\usage{
plot_bestModelConfusionMatrix(trg, bestModelsList, balancedList)
}
\arguments{
\item{trg}{is thetarget}

\item{bestModelsList}{is the list of the best model for each target}

\item{balancedList}{is the list of train (balanced) and test for target}
}
\value{
confusionMatrix object
}
\description{
plot the ConfusionMatrix
}
