% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EyeMetGen_Functions.R
\name{plot_bestModelROC}
\alias{plot_bestModelROC}
\title{plot_bestModelROC}
\usage{
plot_bestModelROC(trg, bestModelsList, balancedList)
}
\arguments{
\item{trg}{is thetarget}

\item{bestModelsList}{is the list of the best model for each target}

\item{balancedList}{is the list of train set (balanced) and test set for the target}
}
\value{
roc object
}
\description{
get the ROC curve
}
