% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EyeMetGen_Functions.R
\name{preprocess_split_dataset}
\alias{preprocess_split_dataset}
\title{preprocess_split_dataset}
\usage{
preprocess_split_dataset(data_in_scope, sd = 2023)
}
\arguments{
\item{data_in_scope}{is the subset with respect the target}
}
\value{
list
}
\description{
generate train and test for the target subset
}
