% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EyeMetGen_Functions.R
\name{preprocessing_recipes}
\alias{preprocessing_recipes}
\title{preprocessing_recipes}
\usage{
preprocessing_recipes(data_in_scope_train, sd = 2020)
}
\arguments{
\item{data_in_scope_train}{is the train dataset}

\item{sd}{is the seed}
}
\value{
dataframe
}
\description{
Pre-Processing the data with{recipes}
}
