% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildModels_Functions.R
\name{random_forest_model}
\alias{random_forest_model}
\title{random_forest_model}
\usage{
random_forest_model(train_data, ctrl, sd = 2023)
}
\arguments{
\item{train_data}{is train set}

\item{ctrl}{is la list (i.e. returned by TrainControl function)}

\item{sd}{is the seed}
}
\value{
train object
}
\description{
Create a caret model
}
