% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Feature_Selection_Functions.R
\name{RFE_FeatureSelection}
\alias{RFE_FeatureSelection}
\title{RFE_FeatureSelection}
\usage{
RFE_FeatureSelection(train_data, test_data, trg, sd = 2023)
}
\arguments{
\item{train_data}{is train set}

\item{test_data}{is test set}

\item{trg}{is target (i.e., outcome)}

\item{sd}{is the seed}
}
\value{
list
}
\description{
Feature Selection - Recursive Feature Elimination
}
