% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EyeMetGen_Functions.R
\name{save_plt}
\alias{save_plt}
\title{save_plt}
\usage{
save_plt(plt, path, dpi = 300, height = 45, width = 45)
}
\arguments{
\item{plt}{ggplot output}

\item{dpi}{of the plot}

\item{height}{height of ggplot output}

\item{width}{width of ggplot output}

\item{model_fitename}{name of model_fite}
}
\description{
save roc curve
}
