% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EyeMetGen_Functions.R
\name{test_roc}
\alias{test_roc}
\title{test_roc}
\usage{
test_roc(model, data)
}
\arguments{
\item{model}{is the train model}

\item{test_data}{are the test data dataframe}
}
\description{
extract AUC from the caret model object
}
