% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EyeMetGen_Functions.R
\name{wilcox.test.bacteria}
\alias{wilcox.test.bacteria}
\title{wilcox.test.bacteria}
\usage{
\method{wilcox.test}{bacteria}(tar.bact, target)
}
\arguments{
\item{tar.bact}{is the full dataset (target+bacteria)}

\item{target}{is the target dataset}
}
\value{
list
}
\description{
wilcox test
}
