% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EyeMetGen_Functions.R
\name{wilcox_test_bacteria}
\alias{wilcox_test_bacteria}
\title{wilcox.test.bacteria}
\usage{
wilcox_test_bacteria(tar.bact, target)
}
\arguments{
\item{tar.bact}{is the full dataset (target+bacteria)}

\item{target}{is the target dataset}
}
\value{
dataframe
}
\description{
wilcox test
}
